refactored GridView to avoid repeating centerpoint calculation and rewrote VectorAnnotation to use normalized 0-1 angles rather than discrete steps.

+++ src/vitro/grid/VectorAnnotation.java
@@ -5,14 +5,14 @@ import java.awt.Color;
 import java.util.*;
 
 public class VectorAnnotation implements Annotation {
-	public final Map<Location, Integer> dirs = new HashMap<Location, Integer>();
+	public final Map<Location, Double> dirs = new HashMap<Location, Double>();
 	public final Color color;
 
 	public VectorAnnotation(Color color) {
 		this.color = color;
 	}
 
-	public VectorAnnotation(Color color, Map<Location, Integer> dirs) {
+	public VectorAnnotation(Color color, Map<Location, Double> dirs) {
 		this(color);
 		this.dirs.putAll(dirs);
 	}
